refactored and added success page
