+++ templates/success.pug
@@ -1,0 +1,8 @@
+extends ./layout.pug
+include ./mixins.pug
+block title
+  title Success
+block body
+  h1 Thank you for creating a new username! Please log back in
+  button.button
+    a(href="/") Log back in
